requirements.txt
Django
python-decouple
django-extensions
django-browser-reload
djangorestframework